+++ NEWS.md
@@ -1,5 +1,10 @@
 # User visible changes in `ConjugateGradient.jl`
 
+# Version 0.3.0
+
+- Status returned by `ConjugateGradient.solve!` is an enumeration which can be compared
+  against an integer or another status.
+
 # Version 0.2.0
 
 - Use `NumOptBase` instead of `LazyAlgebra` for basic operations on variables.